desiño de ventana editar, y de perfil cliente
