+++ lib/Router/strings.dart
@@ -5,4 +5,4 @@ const String loginRoute = 'login';
 const String perfilclienteRoute = 'perfilcliente';
 const String homeRoute = 'home';
 const String tomarPedidoRoute = 'tomarPedido';
-const String editar = 'editarcliente';
+const String editarRoute = 'editaruser';
